requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
# sqlite3 is built-in to Python

# LLM and AI dependencies
ollama>=0.1.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
fastmcp>=0.2.0
anthropic>=0.34.0
openai>=1.51.0
boto3>=1.34.0
chromadb>=0.4.0
httpx>=0.25.0
uvicorn>=0.24.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
python-multipart>=0.0.6

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.9.0
flake8>=6.1.0
